requirements.txt
# Core
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
psycopg2-binary==2.9.10
gunicorn==23.0.0

# Data processing
pandas==2.2.3
openpyxl==3.1.5
numpy==2.2.3

# ETL
requests==2.32.3

# Security
python-dotenv==1.0.1
email-validator==2.2.0

# Barcode generation
python-barcode==0.15.1
Pillow==11.1.0

# PDF export
WeasyPrint==63.1

# Testing
pytest==8.3.4
pytest-cov==6.0.0
factory-boy==3.3.1
Faker==33.3.1
